% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{alphaAlong}
\alias{alphaAlong}
\title{alphaAlong}
\usage{
alphaAlong(x, head = 20, skew = -2)
}
\arguments{
\item{x}{}

\item{head}{}

\item{skew}{}
}
\description{
alphaAlong
}
\examples{
 alphaAlong(x =  , head = 20, skew = -2)
}
