% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie.R
\name{make_moviemap}
\alias{make_moviemap}
\title{make_moviemap}
\usage{
make_moviemap(
  dir,
  outnam,
  outdir,
  nframes = nrow(frmss),
  movie_length_secs = 25
)
}
\arguments{
\item{dir}{}

\item{outnam}{}

\item{outdir}{}

\item{nframes}{}

\item{movie_length_secs}{}
}
\description{
make_moviemap
}
\examples{
make_moviemap(dir =  , outnam =  , outdir =  , nframes = nrow(frmss), movie_length_secs = 25)
}
