% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie.R
\name{makefrm}
\alias{makefrm}
\title{makefrm}
\usage{
makefrm(d, bytime = "15 mins", byid = "id")
}
\arguments{
\item{d}{}

\item{bytime}{}

\item{byid}{}
}
\description{
makefrm
}
\examples{
makefrm(d =  , bytime = 15 mins, byid = id)
}
