% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sizeAlong}
\alias{sizeAlong}
\title{sizeAlong}
\usage{
sizeAlong(x, head = 20, to = c(0.1, 2.5))
}
\arguments{
\item{x}{}

\item{head}{}

\item{to}{}
}
\description{
sizeAlong
}
\examples{
sizeAlong(x =  , head = 20, to = c(0.1, 2.5))
}
