% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_points2lines}
\alias{st_points2lines}
\title{st_points2lines}
\usage{
st_points2lines(x, grp, shift_lon = FALSE)
}
\arguments{
\item{x}{}

\item{grp}{}

\item{shift_lon}{}
}
\description{
st_points2lines
}
\examples{
st_points2lines(x =  , grp =  , shift_lon = FALSE)
}
